Address compilation errors on Windows
It seems that the compiler used by conda-build (a variant of Mingw-64, gcc 5.3.0) is unable to find the isnan() macro from the C math.h library. This might or might not be caused by the #include <algorithm>, which could cause <cmath> to be included, which in turn undefines several C-style macros from the global namespace.